refactor: 카드 브랜드 식별 및 포맷팅 로직 관심사 분리
- 각각을 파일로 분리 -> 재사용성 높이기 및 관심사 분리
- 여러 파일에서 사용하는 카드 브랜드 타입 따로 types 폴더 내부로 분리
