+++ hooks/src/types/cardBrand.ts
@@ -1,0 +1,2 @@
+export type CardBrand = "visa" | "masterCard" | "diners" | "amex" | "unionPay";
+export type CardBrandType = CardBrand | null;
